move service api types into separate project
