+++ module/client/build.gradle.kts
@@ -8,7 +8,7 @@ dependencies {
     annotationProcessor(libs.immutables.value)
 
     api(project(":api"))
-    api(project(":service"))
+    api(project(":service:api"))
     api(libs.dagger.dagger)
     api(libs.immutables.annotations)
     api(libs.jackson.annotations)
