+++ module/client/src/main/java/org/example/age/module/client/SiteClientsConfig.java
@@ -5,7 +5,7 @@ import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import java.net.URL;
 import org.example.age.api.SiteApi;
 import org.example.age.api.client.AvsApi;
-import org.example.age.service.ConfigStyle;
+import org.example.age.service.api.config.ConfigStyle;
 import org.immutables.value.Value;
 
 /** Configuration for the clients used by the service implementation of {@link SiteApi}. */
