+++ module/request-demo/src/main/java/org/example/age/module/request/demo/DemoAccountIdContext.java
@@ -5,8 +5,8 @@ import jakarta.inject.Singleton;
 import jakarta.ws.rs.NotAuthorizedException;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CompletionStage;
-import org.example.age.service.RequestContextProvider;
 import org.example.age.service.api.request.AccountIdContext;
+import org.example.age.service.api.request.RequestContextProvider;
 
 /** Demo implementation of {@link AccountIdContext}. Uses an {@code Account-Id} header. */
 @Singleton
