+++ service/api/src/main/java/org/example/age/service/api/request/RequestContextFilter.java
@@ -1,4 +1,4 @@
-package org.example.age.service;
+package org.example.age.service.api.request;
 
 import jakarta.inject.Inject;
 import jakarta.inject.Singleton;
